Raster and PSTH

Raster plot and PSTH are very useful for neural activities to periodic stimuli. Patterns of firing can be easily seen and further analyized when spikes are aligned with the onset of the stimuli.

Getting Rasters

Rasters will be created as Vector{Vector{T}}, where each subvector will be spike times relative to the stimulus onset. Besides the spike train, you should at least have the onset time of the stimulus (marker::AbstractVector), and specify the length of time before the onset and duration of the stimulus and time after the stimulus.

julia> spike_train = rand(1000) * 50
julia> marker = 0:10:40
julia> raster = SpikeTriggered.Stats.spike_raster(spike_train, marker;
                                           head=0.5, duration=1.0, tail=0.5)
...

Getting PSTH

Using Histogram

Just like get a histogram from any other data. You could either define the bin size of your histogram:

julia> bin_size = 0.5
julia> psth = SpikeTriggered.Stats.spike_histogram(spike_train/raster, bin_size)

Or providing the edges of your bins:

julia> edges = 0:0.01:50.0
julia> psth = SpikeTriggered.Stats.spike_histogram(spike_train/raster, edges)

The latter is preferred, as you could have finer control and you could have a nonlinear timescale. Just don't forget to append the last right edge.

Using Filter

Sometimes, it's arbitary to find a good bin size, or you want to smooth the histogram, you can use a filter to generate a smooth PSTH. Essentially, it is a convolution over time steps provided by projection::AbstractVector.

\[\text{PSTH}(t) = (h * s)(t) = \sum_i \delta(t - t'_i) h (t - t'_i),\; t \in \text{projection}\]

where h is the kernel function, t'_i is the spike time.

Gaussian kernel is mostly common one, so you can generate it by:

julia> psth = SpikeTriggered.Stats.spike_filter_gaussian(spike_train/raster, projection; σ=0.010)

Or, if you want, you can also create your own kernel:

julia> my_kernel(dt; thresh) = dt < thresh ? 0 : 1
julia> psth = SpikeTriggered.Stats.spike_filter(spike_train/raster, projection, my_kernel; thresh=0.5)

Functions

SpikeTriggered.Stats.spike_raster — Function

spike_raster(spk, markers; head, duration, tail) -> Vector{Vector{T}}

Create rasters form spike train and onset time of stimulus.

Arguments:

spk::AbstractVector{T}: spike train

markers::AbstractVector: onset time of stimulus

Keyword Arguments:

head: time included before the event markers

duration: event length

tail: time included after the event duration

offset: flag to offset all the spike time to the onset time. [default: true`]

Return

rasters as Vector{Vector{T}}

SpikeTriggered.Stats.spike_histogram — Function

spike_histogram(spk::AbstractVector, edges::AbstractVector)

PSTH as histogram, where edges are defined by user.

Arguments:

spk::AbstractVector{T}: spike train

edges::AbstractVector: edges of histogram, noted that it should be length of n+1.

Returns:

psth as Vector{T} of length n.

spike_histogram(spk::AbstractVector, binsize::Real)

PSTH as histogram, where edges are created based on the provided binsize.

Arguments:

spk::AbstractVector{T}: spike train

binsize::Real: binsize of histogram. edges will be created as: range(_min_time, step=binsize, length=_L+1)

spike_histogram(raster::Vector{Vector{T}}, args...; norm=true)

PSTH as histogram from averaging trials.

SpikeTriggered.Stats.spike_filter — Function

spike_filter(spk, proj, kernel::Function; kwargs...) where {T <: Real} -> Vector{T}

generating smoothed curve from spike trains. equivalent to convolution.

\[\text{PSTH}(t) = (h * s)(t) = \sum_i \delta(t - t'_i) h (t - t'_i),\; t \in \text{proj}\]

Arguments:

spk: spike train vector

proj: range or vector of timestamps interested

kernel: function of the smoothing kernel, should be (::T; kwargs...)::T

Returns:

psth as Vector{T}, same length as proj.

spike_filter(raster, proj, kernel::Function; kwargs...) where {T <: Real} -> Vector{T}

generating smoothed curve from rasters.

SpikeTriggered.Stats.spike_filter_gaussian — Function

spike_filter_gaussian(spk_or_raster, proj; σ=0.010)

Spike filter using Gaussian function.

\[g(x) = \frac{\exp(-\frac{x^2}{2 σ^2})}{\sqrt{2\pi} σ}\]

SpikeTriggered.PSTH.psth2timestamps — Function

psth2timestamps(psth::Vector{T}; width=1, sporadic=false, upper=Inf) where {T <: Integer} -> Vector{Float64}

convert PSTH to timestamps.

Arguments

psth::Vector{T <: Integer}

Keyword Arguments

width: define the bin width [default: 1]

sporadic: if true, the generation would be random (using rand); otherwise, evenly divided [default: false]

upper: cutoff of the PSTH bin value [default: Inf]

Return

Vector{Float64}

SpikeTriggered.PSTH.psth2trace — Function

psth2trace(hist::Vector{T}, roi; σ=0.03, freq=35) where {T} -> Vector{T}

convert histgram into trace-like data, with a Gaussian filter.

Arguments

hist::Vector{T}: histogram vector

roi: range or a vector of timepoints.

Keyword Arguments:

σ: sigma of the gaussian kernel (in seconds) [default: 0.03]

freq: equivalent sampling frequency (Hz) of the histogram. [default: 35]

Returns:

Vector{T}: with the same length as the roi.