Raster and PSTH

Raster plot and peristimulus time histogram(PSTH) are very useful for neural activities to periodic stimuli or repeated trials. Patterns of firing can be easily seen and further analyized when spikes are aligned with the onsets of the stimulus.

In most cases, spike trains are stored as Vector{T} where T <: Real. Mathematically, it can be represented as series of Dirac delta function:

\[ \operatorname{s}(t) = \sum_i \delta(t - t'_i)\]

where $t'_i$ is the spike times in the spike train vector.

Getting Rasters

Rasters will be created as Vector{Vector{T}} using spike_raster, where each subvector will be spike times relative to the stimulus onset. Besides the spike train, you should at least have the onset time of the stimulus (marker::AbstractVector), and specify the length of time before the onset (head) and duration of the stimulus and time after the stimulus (tail).

using SpikeTriggered
spike_train = rand(1000) * 50
marker_onset = 0:10:40
raster = spike_raster(spike_train, marker_onset; head=0.5, duration=1.0, tail=0.5)
...

Getting PSTHs

Histogram style PSTHs can be generated with spike_histogram, and FHist is used under the hook. In order to have finer control and a nonlinear timescale (e.g., when you want to group each trial into 0-100ms and 100ms-500ms as triggering phase and steady phase), edges of the histogram is taken. So, don't forget to append the last right edge.

edges = 0:0.01:50.0
psth = spike_histogram(spike_or_raster, edges)

Note

If you need PSTHs for each single trial, you can organize the edges into 2d Matrix as [edges x nTrials], which results into 2d raster as [n x nTrials]; an easy way to generate such matrix:

projection = 0:0.001:2
trial_onset = [0,5,10,15,20]
edge_2d = projection .+ trial_onset' # 2d edge matrix
psth_2d = spike_histogram(spike_train, edge_2d)

Note

the PSTH from raster has an additional kwargs norm=true which will normalized the PSTH by the number of trials.

Getting smoothed PSTHs

Sometimes, it's arbitary to find a good bin size, or you just want to smooth the histogram; you can use a kernel function to generate a smooth PSTH using spike_histogram_smoothed. Essentially, it is a convolution over time steps provided by projection::AbstractVector.

\[\operatorname{PSTH}(t) = (\operatorname{h} * \operatorname{s})(t) = \sum_i \delta(t - t'_i) h(t - t'_i),\ t \in \text{projection}\]

where h is the kernel function, $t'_i$ is the spike time. By default, it uses a Gaussian kernel function (Stats.gaussian_kernel).

# Gaussian kernel with default sigma = 0.005
spike_histogram_smoothed(spike_or_raster, projection)

# change the Gaussian kernel width
spike_histogram_smoothed(spike_or_raster, projection; σ=0.010)

Note

The kernel function shoudl be defined as (dt; kwargs...) -> ...:

my_kernel(dt; thresh) = dt < thresh ? 0 : 1
psth = spike_histogram_smoothed(spike_or_raster, projection, my_kernel; thresh=0.5)

Note

There is also a norm=true flag, to normalize the PSTH by its maximum value, so the PSTH will be between [0, 1].

Functions

exported functions

SpikeTriggered.Stats.spike_raster — Function

spike_raster(spike_train::AbstractVector{T}, markers::AbstractVector; head=0.5, duration=1.0, tail=0.5, offset=true) -> Vector{Vector{T}}

Create rasters from spike_train and onset times of stimulus (markers), with interval of (marker-head, marker+duration+tail].

If offset is true, spike times of each trial be relative to the onset time.

SpikeTriggered.Stats.spike_histogram — Function

spike_histogram(spike_train::AbstractVector{T}, edges::AbstractVector; dtype::Type, kwargs...) -> Vector{T}

Peri-stimulus time histogram. edges should be length of n+1 to create a histogram of length n.

If spike_train is empty, it will return zeros(T, n)

If counttype is not specified, the returned vector will have same type T

spike_histogram(spike_train, edges::AbstractMatrix; kwargs...) -> Matrix{T}

Peri-stimulus time histogram with multiple trials. edges should be shape of [n+1 x nRepeats] to create a histogram of shape [n x nRepeats].

If dims is specified, trial dimension will be overwrite.

Additional kwargs will be passed to spike_histogram.

spike_histogram(raster::SpikeRaster{T}, args...; norm=true, kwargs...) -> VecOrMat{T}

PSTH from averaging rasters.

If norm is true, PSTH will be normalized by the length of trials.

SpikeTriggered.Stats.spike_histogram_smoothed — Function

spike_histogram_smoothed(spike_train::SpikeTrain{T}, projection::AbstractArray, kernel::Function=gaussian_kernel; norm=true, kwargs...) -> Vector{T}

Generating smoothed curve from spike trains. Equivalent to convolution.

\[\text{PSTH}(t) = (h * s)(t) = \sum_i \delta(t - t'_i) h (t - t'_i),\; t \in \text{proj}\]

If norm is true, the results will be normalized by the maximum value.

All kwargs will be passed to kernel function.

spike_histogram_smoothed(raster, args...; kwargs...) where {T <: Real} -> Vector{T}

Generating smoothed PSTH from rasters.

If norm is true, PSTH will be normalized to the maximum number; otherwise, it will be divided by the number of trials in the raster.

utility functions

SpikeTriggered.Stats.gaussian_kernel — Function

gaussian_kernel(x::T; σ::T=0.005)

Inline Gaussian function.

\[g(x) = \frac{\exp(-\frac{x^2}{2 σ^2})}{\sqrt{2\pi} σ}\]

SpikeTriggered.Stats.get_histogram_center — Function

get_histogram_center(edges)

get the center of edge vector. make sure edges are sorted.

SpikeTriggered.Stats.histogram_fhist — Function

histogram_fhist(u_arr, edges; kwargs...) -> Vector

Count histogram using FHist.Hist1D, which has similar performance as GSL.

kwargs will be relayed to FHist.Hist1D; for example, one could set element type of the histogram to Float32 by passing counttype=Float32.

SpikeTriggered.Stats.histogram_gsl — Function

histogram_gsl(u_arr, edges) where {T <: Real}

Count histogram by providing the edges. If provides n+1 edges, it would return n length histogram. Edges would work as: left bound <= value < right bound.

Arguments

u_arr: a vector of values

edges: a vector of edges

Note: StatsBase.Histogram is slower than GSL but with less memory footprint.