Reverse Correlation

Probabily the most important analysis, the foundation of estimating the spatiotemporal receptive fields, of the first-order component (Spike-Triggered Average) and higher-order component (Spike-Triggered Covariance)[schwartzSpiketriggeredNeuralCharacterization2006]. The basic idea is the features of the stimulus that drive the cellular responses (e.g., spikes) is the features of the receptive fields.

Usually the features here are continous random vairables (e.g., noise stimuli), it can also be the movie stimulus itself. No matter what feature you choose, the stimulus matrix for this analysis should be organized as [nDimensions x nTimepoints]; and the reverse correlation would be [nDimensions x nFrames].

Spike-Triggered Average

The first order component, which is equivalent to the weighted average of the stimulus before spikes.

\[\operatorname{STA}(n) = \frac{1}{N} \sum_{i=1}^N \vec{s}(n, t_i)\]

where $t_i$ is the spike time and $\vec{s}(n, t_i)$ is the stimulus feature vector related to the spike.

bootstrapped spike-tiriggered average

TODO

Spike-Triggered Covariance

TODO

References

Functions

exported functions

SpikeTriggered.Stats.spike_triggered_average — Function

spike_triggered_average(stimulus, psth; n=10)
spike_triggered_average(stimulus, spike_train, marker; n=10)
spike_triggered_average(stimulus, raster, marker; n=10)

Get the STA, which can be reshaped into 2d/3d matrix with make_strf and hstack_strf.

If psth is 2d matrix (nTimepoints x nRepeats), the average of PSTH across trials will be used.

\[A = \frac{1}{N} \sum_{n=1}^{N}\vec{s}(t_n)\]

schwartzSpiketriggeredNeuralCharacterization2006Schwartz, O., Pillow, J. W., Rust, N. C. & Simoncelli, E. P. Spike-triggered neural characterization. Journal of Vision 6, 484–507 (2006).