APIs

Stats

SpikeTriggered.Stats.benjamini_hochberg_qvalue — Method

benjamini_hochberg_qvalue(pvalue; C=1) -> (q::Array, k::Array)

False discovery rate test for any list of pvalues. This function will return the q value and the corresponding k index.

\[P_k \leq \frac{k}{N} \alpha\]

\[q_k = \frac{N}{k} P_k \leq \alpha\]

k is generated from invperm(sortperm(...)).

SpikeTriggered.Stats.burst_detect — Method

burst_detect(spk::Vector; t_silence=0.07, t_isi=0.03, nofs=nothing, keep_index=false) -> Vector{Vector{T}}

Burst firing identification based on the rules:

a burst would followed/preceded by a silence t_silence time;

if t_silence is a 2 element tuple, the t_silence[1] will set the preceded time and t_silence[2] will be the followed time.

each spike spaced at most t_isi time, otherwise terminated

(optional) at least nofs[1] spikes within the first nofs[2] time

If keep_index is true, will return indices of spikes within the input; otherwise, the spike times will be returned.

References:

Vaingankar et al 2012

SpikeTriggered.Stats.burst_detect_lgn — Method

detect_burst_lgn(spk; kwargs...)

Bursts were defined as two or more spikes, each spaced ≤4 ms apart following ≥100 ms of silence, bursts rarely lasted more than 10 ms for LGN.

SpikeTriggered.Stats.burst_detect_trn — Method

detect_burst_trn(spk; kwargs...)

≥5 spikes within 70 ms, spaced ≤30 ms apart following ≥70 ms of silence; bursts were terminated when the interspike interval exceeded 30 ms, Figure 1B. Thus defined, typical bursts had 5–17 spikes. Typical bursts lasted between 70 and 100 ms.

SpikeTriggered.Stats.burst_interpolate — Method

interp_burst(burst_iti::AbstractVector; n=13, interp_t=:steffen) -> Vector{Float64}
interp_burst(bursts::Vector{Vector{Real}}; n=13, interp_t=:steffen) -> Matrix{Float64} [n x nBursts]

Make interpolation of burst inter-spike-interval to generate burst patterns.

n set the interpolation bin length. interp_t set the type of interpolation, currently only supports [:linear, :polynomial, :cspline, :steffen]. The results will always be in Cdouble precision.

SpikeTriggered.Stats.covariogram — Method

The covariogram is defined as:

\[V \equiv \langle (S_1^{r} - P_1) \odot (S_2^{r} - P_2) \rangle = \langle S_1^{r} \odot S_2^{r} \rangle - P_1 \odot P_2\]

SpikeTriggered.Stats.dispersion_index — Method

dispersion_index(psth; corrected=true, kwargs...) -> Real

Also known as the Fano factor.

\[D = \frac{\sigma^2}{\mu}\]

optional arguments:

dims: calculate the index over dimension dims.

Check out more on wikipedia.

SpikeTriggered.Stats.gaussian_kernel — Method

gaussian_kernel(x::T; σ::T=0.005)

Inline Gaussian function.

\[g(x) = \frac{\exp(-\frac{x^2}{2 σ^2})}{\sqrt{2\pi} σ}\]

SpikeTriggered.Stats.get_footprint_map — Method

get_footprint_map([func_pval,] raw::AbstractArray; alpha=0.01, fdr_c=1, collapse=false, gridsize=nothing)

Create a footprint map based on the q-values from func_pval(raw) using Benjamini–Hochberg procedure; alpha as the threshold, fdr_c as the constant.

\[P_k \leq \frac{k}{N \cdot C} \alpha\]

For arbiturary join distribution of p-values, one could set fdr_c as ln(N)+1/N+γ (Euler constant). (using benjamini_hochberg_constant(N))

If collapse is true and gridsize is specified, a 2d time-collapsed footprint map will be returned.

color codes as:

1: ON response

0: NO response

-1: OFF response

SpikeTriggered.Stats.get_footprint_mask — Method

get_footprint_mask([func_pval::Function,] raw::AbstractArray; alpha=0.01, fdr_c=1)

Create a binary footprint mask based with Benjamini–Hochberg procedure using arbiturary q-values converted from raw statistic values by func_pval.

In most cases, one should use get_footprint_map which can also generate 3d mask with addtional polarity information.

If func_pval is not set, zscore_pvalue will be used, which assumes inputs are zscore from standard normal distribution.

alpha if the threshold for false discovery rate, default as 0.01.

fdr_c is the constant value for correcting Benjamini–Hochberg procedure with joint distributions among pvalues.

SpikeTriggered.Stats.get_footprint_mask_from_qvalue — Method

get_footprint_mask_from_qvalue(qval, k; alpha=0.01)

create a binary footprint mask based on the q-values and the corresponding permutation vector. Using Benjamini–Hochberg procedure.

SpikeTriggered.Stats.get_histogram_center — Method

get_histogram_center(edges)

get the center of edge vector. make sure edges are sorted.

SpikeTriggered.Stats.histogram_fhist — Method

histogram_fhist(u_arr, edges; kwargs...) -> Vector

Count histogram using FHist.Hist1D, which has similar performance as GSL.

kwargs will be relayed to FHist.Hist1D; for example, one could set element type of the histogram to Float32 by passing counttype=Float32.

SpikeTriggered.Stats.histogram_gsl — Method

histogram_gsl(u_arr, edges) where {T <: Real}

Count histogram by providing the edges. If provides n+1 edges, it would return n length histogram. Edges would work as: left bound <= value < right bound.

Arguments

u_arr: a vector of values

edges: a vector of edges

Note: StatsBase.Histogram is slower than GSL but with less memory footprint.

SpikeTriggered.Stats.hstack_strf — Method

hstack_strf(val; gridsize) -> Array{T, 2} Sx(SxT)

convert strf 3d matrix into horizontal matrix.

When plotting with heatmap, REMEMBER to use the transversed matrix and reversed y axis.

SpikeTriggered.Stats.image2dataset — Method

image2dataset(image::AbstractMatrix) -> ([X Y], [Z])

convert 2d image into (x, y) coordinates and vector of elements (z). Primarilly used for 2d gaussian fit.

SpikeTriggered.Stats.make_strf — Method

make_strf(val; gridsize) -> Array{T, 3} SxSxT

convert strf vector into 3d spatiotemporal matrix.

When plotting with heatmap, REMEMBER to use the transversed matrix and reversed y axis.

SpikeTriggered.Stats.schreiber_score_matrix — Method

Schreiber reliability score:

\[R_{\text{corr}} = \frac{2}{N(N-1)} \sum_{i = 1}^N \sum_{j = i+1} ^N \frac{\vec{s}_i \cdot \vec{s}_j}{|\vec{s}_i| |\vec{s}_j|}\]

Input in shape of [nPSTH x nTrial]

reference: dx.doi.org/10.1016/S0925-2312(02)00838-X

SpikeTriggered.Stats.spike_histogram — Method

spike_histogram(spike_train, edges::AbstractMatrix; kwargs...) -> Matrix{T}

Peri-stimulus time histogram with multiple trials. edges should be shape of [n+1 x nRepeats] to create a histogram of shape [n x nRepeats].

If dims is specified, trial dimension will be overwrite.

Additional kwargs will be passed to spike_histogram.

SpikeTriggered.Stats.spike_histogram — Method

spike_histogram(spike_train::AbstractVector{T}, edges::AbstractVector; dtype::Type, kwargs...) -> Vector{T}

Peri-stimulus time histogram. edges should be length of n+1 to create a histogram of length n.

If spike_train is empty, it will return zeros(T, n)

If counttype is not specified, the returned vector will have same type T

SpikeTriggered.Stats.spike_histogram — Method

spike_histogram(raster::SpikeRaster{T}, args...; norm=true, kwargs...) -> VecOrMat{T}

PSTH from averaging rasters.

If norm is true, PSTH will be normalized by the length of trials.

SpikeTriggered.Stats.spike_histogram_smoothed — Method

spike_histogram_smoothed(spike_train::SpikeTrain{T}, projection::AbstractArray, kernel::Function=gaussian_kernel; norm=true, kwargs...) -> Vector{T}

Generating smoothed curve from spike trains. Equivalent to convolution.

\[\text{PSTH}(t) = (h * s)(t) = \sum_i \delta(t - t'_i) h (t - t'_i),\; t \in \text{proj}\]

If norm is true, the results will be normalized by the maximum value.

All kwargs will be passed to kernel function.

SpikeTriggered.Stats.spike_histogram_smoothed — Method

spike_histogram_smoothed(raster, args...; kwargs...) where {T <: Real} -> Vector{T}

Generating smoothed PSTH from rasters.

If norm is true, PSTH will be normalized to the maximum number; otherwise, it will be divided by the number of trials in the raster.

SpikeTriggered.Stats.spike_raster — Method

spike_raster(spike_train::AbstractVector{T}, markers::AbstractVector; head=0.5, duration=1.0, tail=0.5, offset=true) -> Vector{Vector{T}}

Create rasters from spike_train and onset times of stimulus (markers), with interval of (marker-head, marker+duration+tail].

If offset is true, spike times of each trial be relative to the onset time.

SpikeTriggered.Stats.spike_train_spectrum_power — Method

spike_train_spectrum_power(spks::AbstractSpikeTrain{T}, freq; t_len=1) where {T <: AbstractFloat}

The spike-train power spectrum is computed from the Fourier transform of the spike train, as in (Dummer, et al., 2014).

\[S(f) = \frac{\langle \tilde{x} \tilde{x}^\ast \rangle}{T}\]

The neural spike train is represented by a sum of delta functions at the spike times:

\[x(t) = \sum_i \delta(t - t_i)\]

The Fourier transform for the time window is defined by

\[\tilde{x}(f) = \int_0^{T} e^{- 2 \pi j f t} x(t)\; dt\]

Reference:

Dummer, B., Wieland, S. & Lindner, B. Self-consistent determination of the spike-train power spectrum in a neural network with sparse connectivity. Front. Comput. Neurosci. 8, (2014). DOI: 10.3389/fncom.2014.00104

SpikeTriggered.Stats.spike_triggered_average — Method

spike_triggered_average(stimulus, psth; n=10)
spike_triggered_average(stimulus, spike_train, marker; n=10)
spike_triggered_average(stimulus, raster, marker; n=10)

Get the STA, which can be reshaped into 2d/3d matrix with make_strf and hstack_strf.

If psth is 2d matrix (nTimepoints x nRepeats), the average of PSTH across trials will be used.

\[A = \frac{1}{N} \sum_{n=1}^{N}\vec{s}(t_n)\]

SpikeTriggered.Stats.spike_triggered_average_suite — Method

spike_triggered_average_suite(args...; kwargs...) -> (; diff, on, off)

Create STAs of difference map, on map and off map; returns a named tuple.

SpikeTriggered.Stats.spike_triggered_average_zscore — Method

spike_triggered_average_zscore(args...; kwargs...)

Estimate z-scores of STA.

Same args... as spike_triggered_average.

If n is specified, only n frames are included for zscore (same as sta, default as 10`).

bootstrap set the bootstrapping iterations, default as 1000.

SpikeTriggered.Stats.spike_triggered_covariance — Method

spike_triggered_covariance(X::AbstractMatrix{T}, y::AbstractArray{T}; n::Integer=10, verbose=true) where {T, U} -> Matrix{T}

SpikeTriggered.Stats.spike_xcorr — Method

spike_xcorr(target, reference, roi) where {T <: Real}

The crosscorrelogram shows a count of the spikes of the target cell at specific time delays with respect the spikes of the reference cell.

If there are multiple trials, use the spike_xcorr_shifted to correct any bias with a shift predictor.

SpikeTriggered.Stats.spike_xcorr_shifted — Method

spike_xcorr_shifted(target, reference, roi; shift_t, shift_t_end) where {T <: Real}

The crosscorrelogram shows a count of the spikes of the target cell at specific time delays with respect the spikes of the reference cell. Corrected with a shift predictor.

Keyword Arguments:

shift_t: Δt of each shift, usually shift the spike by trials

shifttend: the end limit of spike_time, usually the end time of the last trial. [default: maximum(reference)]

SpikeTriggered.Stats.split_tonic_burst — Method

split_tonic_burst(spk; detector, kwargs...) -> (; burst::Vector{Vector}, tonic::Vector)

Split spike train into burst and tonic groups using detector function.

SpikeTriggered.Stats.split_tonic_cardinal — Method

split_tonic_cardinal(spk; detector, kwargs...) -> (; burst::Vector, tonic::Vector)

Split spike train into burst and tonic groups using detector function. But only the cardinal spike of burst is returned.

SpikeTriggered.Stats.zscore_pvalue — Method

zscore_pvalue(zscore::AbstractArray; two_tailed=true)

get pvalue from standard normal distribution

SpikeTriggered.Stats.SpatialRF.schiller_overlap_index — Method

schiller_overlap_index(e1::GaussianEllipse, e2::GaussianEllipse)

overlap of two spatial receptive fields based on Schiller's overlap index.

For two receptive fields with the Gaussian parameters ($\alpha$, $\sigma_a$, $\sigma_b$, $\theta$, $x_0$, $y_0$), the 2D overlap index is defined as:

\[\Omega = \frac{\sigma_{12} + \sigma_{21} - d}{\sigma_{12} + \sigma_{21} + d}\]

Where $d$ is the Euclidian distance between the centers, and:

\[\sigma_{ij} = \frac{\sigma_{ai}\sigma_{bi}}{\sqrt{ \sigma_{ai}^2 \sin^2(\theta_i - \phi_{ij}) + \sigma_{bi}^2 \cos^2(\theta_i - \phi_{ij}) }}\]

\[\phi_{ij} = \arctan\left(\frac{y_{0j} - y_{0i}}{x_{0j} - x_{0i}}\right)\]

References

Schiller et al., 1979

Wang et al., 2007

Waveforms

SpikeTriggered.Waveforms.detect_cross_pnt — Method

detect_cross_pnt(arr; thresh, gap=1, dir=:up, ref=nothing)

detect data rise or data down points.

Arguments

arr: a vector of traces.

thresh: threshold

SpikeTriggered.Waveforms.gaussian_filter_gsl — Method

gaussian_filter_gsl(arr::Vector{Float64}; K::Integer, α::Real=1)

The GSL gaussian kernel is defined by:

\[G(k) = \exp\left(- \frac{k^2}{2 \sigma^2}\right)\]

for $-(K-1)/2 \leq k \leq (K-1)/2$, and $K$ is the size of the kernel. The parameter $\alpha$ specifies the number of standard deviations $\sigma$ desired in the kernel.

\[\sigma = \frac{K-1}{2 \alpha}\]

Stimulus

SpikeTriggered.Stimulus.apply_video_nonlinearity — Function

apply_video_nonlinearity(video, bias=0) -> video

For sparse noise stimuli, to retrieve nonlinear map (ON map and OFF map) by ignore bright/dark spots.

Arguments

video: the original video matrix

bias: could be either values or symbols

ON: :on or -1

OFF: :off or 1

DIFF: :diff or 0

SpikeTriggered.Stimulus.load_snra — Method

load_snra(snf_path::String) -> Vector{Int16}

read snf sparse noise pseudorandom files

Synopsis

struct SNF {
    int16_t n_order;
    int32_t base_poly; // not sure what it represents
    int16_t* snra; // the length would be 2^n_order
};

SpikeTriggered.Stimulus.parse_snra — Method

parse_snra(snra::Array{T}, grid_size::Integer) -> NamedTuple(:ids, :row, :col, :sign)

parse the snra array to sparse noise image info.